DBC22-2524: saved cams and routes now update on every refresh
