+++ src/frontend/src/Components/shared/header/UserNavigation.js
@@ -1,10 +1,9 @@
 // React
-import React, { useCallback, useContext, useEffect } from 'react';
+import React, { useCallback, useContext } from 'react';
 
 // Redux
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { memoize } from 'proxy-memoize';
-import { updateFavCams, updateFavRoutes } from '../../../slices/userSlice';
 
 // External imports
 import { DropdownButton } from 'react-bootstrap';
@@ -21,8 +20,6 @@ import { useMediaQuery } from '@uidotdev/usehooks';
 
 // Internal imports
 import { AuthContext } from "../../../App";
-import { getFavoriteCameraIds } from '../../data/webcams';
-import { getFavoriteRoutes } from '../../data/routes';
 
 // Styling
 import './UserNavigation.scss';
@@ -32,35 +29,10 @@ export default function UserNavigation(props) {
   const { authContext, setAuthContext } = useContext(AuthContext);
 
   // Redux
-  const dispatch = useDispatch();
   const { favCams, favRoutes } = useSelector(useCallback(memoize(state => ({
     favCams: state.user.favCams,
     favRoutes: state.user.favRoutes
   }))));
-
-  /* data hooks and functions */
-  const initCams = async () => {
-    // Get saved cam ids and map into a list of integers
-    const favCamsData = favCams ? favCams : await getFavoriteCameraIds();
-    if (favCams === null) {
-      const favCamIds = favCamsData.map(webcam => webcam.webcam);
-      dispatch(updateFavCams(favCamIds));
-    }
-  }
-
-  const initRoutes = async () => {
-    const favRoutesData = favRoutes ? favRoutes : await getFavoriteRoutes();
-    if (favRoutes === null) {
-      dispatch(updateFavRoutes(favRoutesData));
-    }
-  }
-
-  useEffect(() => {
-    if (authContext.loginStateKnown && authContext.username) {
-      initCams();
-      initRoutes();
-    }
-  }, [authContext]);
 
   /* Helpers */
   const toggleAuthModal = (action) => {
